+++ packages/imperative/src/rest/src/session/doc/ISession.ts
@@ -11,6 +11,7 @@
 
 import * as SessConstants from "../SessConstants";
 import { ProxyVariables } from "./ProxyVariables";
+import { IAuthCache } from "./IAuthCache";
 
 /**
  * Session interface for maintaining cookie and protocol information
@@ -179,13 +180,27 @@ export interface ISession {
      * The type in authTypeOrder[0] is used first, authTypeOrder[1] second, etc.
      * Values are specified using SessConstants.AUTH_TYPE_XXX values.
      *
-     * The authTypeOrder property is currently controlled (hard-coded) within Zowe SDK functions.
-     * More control for Zowe consumers is anticipated in the future.
+     * The authTypeOrder property is specified by the end user and is managed by
+     * Zowe SDK functions. Consuming applications should not modify the contents
+     * of the authTypeOrder property.
      *
      * @type {string[]}
      * @memberof ISession
      */
-    authTypeOrder?: string[];
+    authTypeOrder?: SessConstants.AUTH_TYPE_CHOICES[];
+
+    /**
+     * The _authCache property contains a cache of credentials which are available
+     * for authentication. It also contains properties related to whether the user
+     * supplied the authentication order or whether we are using a default order.
+     *
+     * The _authCache property is created and managed by Zowe SDK functions.
+     * Consuming applications should not modify the contents of the _authCache property.
+     *
+     * @type number
+     * @memberof ISession
+     */
+    _authCache?: IAuthCache;
 
     /**
      * Specifies external proxy settings